Add tool calls and make minor refactors
Added tool calling capabilities using standard JSON tool definitions and reworked the agent's code so it uses tool calling to generate the morning wakeup as part of is self task generation.

Also refactored the model and tokenizer to be stored in `Agent` class dicts so that any given model is only loaded in once.