Refactor to separate `Model` from `Agent`.
Made a significant refactor by creating an independent `Model` ABC which is separate from the `Agent` class and its sub-classes. The aim here is to allow me to use a variety of different models by abstracting away the particulars of tool use, reasoning, initialization, etc. I looked into using LangChain or ADK for this, but both are, for the moment, seemingly unsatisfactory. Mainly this is because it's not clear to me what initialization and generation parameters different base models (gpt, clause, etc.) can take and these paradigms obscure this rather than making it clearer.

Also separated out the wakeup message generation to a new `Agent`.